+++ bluetti_mqtt/core/__init__.py
@@ -4,6 +4,7 @@ from .devices.ac300 import AC300
 from .devices.ac500 import AC500
 from .devices.ep500 import EP500
 from .devices.ep500p import EP500P
+from .devices.ep600 import EP600
 from .devices.eb3a import EB3A
 from .commands import DeviceCommand, QueryRangeCommand, UpdateFieldCommand, UpdateRangeCommand
 from .response import CommandResponse